pipeline/requirements.txt
pymupdf==1.24.1
pdf2image==1.17.0
PyPDF2==3.0.1

Pillow==10.0.0
pytesseract==0.3.10

spacy==3.7.2

apache-airflow==2.9.1
apache-airflow-providers-http==4.9.0

python-dotenv==1.0.1
requests==2.32.3
regex==2024.5.15
loguru==0.7.2
python-docx==1.1.0
google-generativeai==0.7.2
